+++ src/Data/Singletons/Syntax.hs
@@ -132,3 +132,30 @@ buildLetDecEnv = go emptyLetDecEnv
     go acc (DInfixD f n : rest) =
       go (infixDecl f n <> acc) rest
     go acc (DPragmaD{} : rest) = go acc rest
+
+-- See Note [Derived Eq instances]
+data DerivedEqDecl = DerivedEqDecl
+  { ded_mb_cxt:: Maybe DCxt
+  , ded_type  :: DType
+  , ded_cons  :: [DCon]
+  }
+
+{- Note [Derived Eq instances]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Most derived instances are handled in Data.Singletons.Partition.partitionDecs.
+There is one notable exception to this rule, however: Eq instances.
+The reason is because deriving Eq in singletons not only derives PEq/SEq instances,
+but it also derives SDecide instances. This additional complication makes Eq
+difficult to integrate with the other deriving machinery, so we handle it
+specially in Data.Singletons.Promote and Data.Singletons.Single
+(depending on the task at hand).
+
+The DerivedEqDecl type encodes just enough information to recreate
+the derived Eq instance:
+
+1. Just the instance context, if it's standalone-derived, or Nothing if it's in
+   a deriving clause (ded_mb_cxt)
+2. The datatype, applied to some number of type arguments, as in the
+   instance declaration (ded_type)
+3. The datatype's constructors (ded_cons)
+-}